requirements.txt
discord.py ~= 2.5.2
python-dotenv ~= 1.1.0
flask ~= 3.1.1